refactor: main에서 parser함수를 CalculatorParser.java로 분리
